refactor filehandler to use base class
