+++ lib/puma/redeploy/file_handler.rb
@@ -5,32 +5,17 @@ require 'forwardable'
 module Puma
   module Redeploy
     # file based redeploy handler
-    class FileHandler
-      extend Forwardable
-      attr_reader :watch_file
-
+    class FileHandler < BaseHandler
       def initialize(watch_file:, deployer:, logger:)
-        @watch_file = watch_file
-        @logger = logger
+        super
         @touched_at = touched_at
-        @deployer = deployer
-      end
-
-      def_delegators :@deployer, :deploy
-      def needs_redeploy?
-        return false unless (mtime = touched_at) != @touched_at
-
-        @touched_at = mtime
-        true
       end
 
       def archive_file
-        File.read(watch_file)&.strip
+        File.read(watch_file).strip
       end
 
       private
-
-      attr_accessor :logger
 
       def touched_at
         if File.exist?(watch_file)
